Fix DNS resolution.
According to [an issue][0] with kubeadm, Ubuntu 18.04 changed its DNS
configuration in a way that broke kube-dns. One user suggested removing
`/etc/resolv.conf` and symlinking it to systemd's as a solution. This
worked when prepended to the user-data, so it should also work as part
of the bootstrap script.

[0]: https://github.com/kubernetes/kubeadm/issues/787