Port component health checks from GCP K8s image.
During presentation on Datadog's usage of K8s[0], they mention that the
official GCP image has a health-checking script for both Docker and
Kubelet. If either time out, it forcibly kills them. This script and
supporting files can be found in the K8s repository. I've ported them to
work with the EKS AMI. For simplicity, I removed support for container
runtimes besides Docker, since we're likely to stick with it.

[0]: https://www.youtube.com/watch?v=2dsCwp_j0yQ
[1]: https://github.com/kubernetes/kubernetes/tree/master/cluster/gce/gci