Switch to Docker 18.06.
May hold off on this--K8s 1.11 is only officially validated against
Docker 17.03. So technically our AMI already uses an unvalidated
version, because 17.06 and 18.06 are the only ones provided officially
in the Ubuntu repos. Once EKS updates to K8s 1.12, though, this branch
should be good to merge, as that version is validated with Docker 18.06.

Update 2019-02-12: yesterday CVE-2019-5736 was announced.[0] Though the
risk of exploitation is small, it still seems prudent to update to a
patched version of Docker. Ubuntu hasn't released one for the 17.12
series, so moving to 18.06 is the path of least resistance. We'll try
this out in the upcoming test cluster.

[0]: https://seclists.org/oss-sec/2019/q1/119