Handle a missing instance_type better in bootstrap
In a strange case that appears to have been a brief metadata server
issue, the "MAX_PODS" selection in bootstrap caused the nodes to fail to
join the cluster.

The log messages were not very helpful, since it triggered a bash math
function error on a string.  The specific failure appears to have been
returning *too much* data, after searching for an empty string matched
every line in the eni-max-pods.txt file.

In the rare case that an instance_type isn't set, search for a known
"unset" value to get predictable behavior.