Add --max-pods param for bootstrap script
Maximum number of pods per node specified at /etc/eks/eni-max-pods.txt
and it is the limit of network interfaces for the instance type. This
number form an amount of memory in Mi kubernetes reserved with formula
11 * $max_num_pods + 255.

Nodes could have a setup when maximum number of pods defined and
it is less than value at the file. Here I add an ability to use
command line value to cover such cases.