Add PRIVATE_BIN_REPO support for pulling binaries
As a EKS ami builder, I want to be able to pull binaries from my own
repo of cached artifacts so that I can ensure availability,
reproducibility, and security.

In my particular case, the binaries are not currently available in
s3://amazon-eks.s3.us-gov-west-1.amazonaws.com and egress rules
are tightly controlled.

How to use:

1. User caches/stages the binaries in their own artifact repo.  At
   present, these are:
     - kubelet
     - aws-iam-authenticator
     - cni-plugins-linux

2. User adds the PRIVATE_BIN_REPO variable to their custom
   eks-worker config

Given that this is an edge uncommon use case, the PRIVATE_BIN_REPO user
variable has not been added to the eks-worker config by default.  Users
will continue to use the existing behavior by default.