Capture the disk usage from pod local storage (#1019)
This captures the storage from the top transient layer of the
overlay filesystem where files that the pod creates while running
are stored.  This allows identifying situations where pods are
writing log files or other data to disk to both identify potential
causes for full disks as well as I/O throttling.