Install containerd/nerdctl
According to the default ctr cli man page bundled with containerd

"ctr is an unsupported debug and administrative client for interacting
with the containerd daemon. Because it is unsupported, the commands,
options, and operations are not guaranteed to be backward compatible
or stable from release to release of the containerd project. "

The nerdctl cli program provided by containerd at
https://github.com/containerd/nerdctl

is meant to be a Docker-compatible CLI for containerd.