Modified check for AWS_ACCESS_KEY_ID
Modified the install-worker.sh script to check for the existence (-v) of the AWS_ACCESS_KEY_ID variable rather than checking whether or not the variable is empty (-n). This modification will allow the installation script to work as a build component in an EC2 Image Builder pipeline. Currently these checks generate "unbound variable" errors.