Add FIPS building support
Addresses #1002.

Description of Changes:

Based on #1028 which was based on #898.

This change adds a new variable called `enable_fips` which will
install openssl and enable fips mode as a kernel paramter on boot.
Additionally fips mode can be enabled while running make by setting
`enable_fips=true` on the command line which will add `-fips` to the
ami name and set `enable_fips` to `true` when building.