Support building for Gov.cloud
Fixes #1536

if a suser only has access to an AWS account in GovCloud then obtaining
the binary versions from S3 will fail.

As this bucket is public we do not need any authentication so add
--no-sign-request  and hardcode the region (which would not be the
default region for any govCloud).

chose `us-west-2` as the region to use as this appears to be the
canonical location.   As it is just metdata using a region on the other
side of the world should not matter too much when the overhead of the
packer build is taken into account