Add ability to specify a user data file
Sometimes we need to execute some actions on the instance along with the
eks build script but currently there is not a way to inject these
additional steps as part of the existing build process. This change adds
the ability to specify `user_data_file` which can be passed to the build
instance on startup.