account for the possibility of there being no aws cli binary
the generate-version-info.sh script presently assumes that the aws
binary will be present. but this isn't always the case depending on the
base AMI being used. allow for the possibility that we will not have an
aws binary so the script doesn't error out.

otherwise we end up with errors like so:

2024-03-08T13:48:36-05:00: ==> amazon-ebs: Provisioning with shell
script:
/home/jo.diaz/projects/amazon-eks-ami/templates/al2/../shared/provisioners/generate-version-info.sh
2024-03-08T13:48:37-05:00:     amazon-ebs:
/home/ec2-user/script_1364.sh: line 26: aws: command not found
2024-03-08T13:48:37-05:00: ==> amazon-ebs: Provisioning step had errors:
Running the cleanup provisioner, if present...
2024-03-08T13:48:37-05:00: ==> amazon-ebs: Terminating the source AWS
instance...