Fix bad pattern for "ignore zero" filter
In `get_io_throttled_processes`, the generated output is filtered to
remove those process that have no throttling. It used a pattern that
matched any throttling value the ended with the digit zero, rather than
a match of the value zero.

This change repairs that. A process with `1000000` centiseconds of delay
will now show up in the output.

Fixes #2255.